requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation
requests
python-dotenv
pydantic-ai-slim[google,mcp]
livekit-agents[tavus]~=1.0
tzdata
flask[async]
flask-cors
livekit-agents[bey]~=1.2
sentence-transformers
torch
tqdm